fix: fix datatypes

+++ STM/max6675.c
@@ -11,11 +11,11 @@ extern SPI_HandleTypeDef hspi1;
 float max6675_readTemperature(void)
 {
     uint8_t rxData[ARRAY_SIZE];
-    int16_t Temperature = 0;
+    float Temperature = 0;
     SPI_CS_SET_LOW();
     HAL_SPI_Receive(&hspi1, rxData, ARRAY_SIZE, SPI_WAIT_TIME);
     SPI_CS_SET_HIGH();
-    Temperature = (rxData[1] | (rxData[0] << BYTE_SIZE)) >> (REGISTER_OFFSET + DEGREES_OFFSET);               // Temperature Data Extraction
+    Temperature = (float) (rxData[1] | (rxData[0] << BYTE_SIZE)) >> (REGISTER_OFFSET + DEGREES_OFFSET);               // Temperature Data Extraction
 //    HAL_Delay(250);                                       // Waits for Chip Ready(according to Datasheet, the max time for conversion is 220ms)
     return Temperature;
 }
